Avoid injecting templating service in constructor
It seems that when using templating service we can ran into a ServiceCircularReferenceException because it "relies on a lot of other services, so it is rather easy to have this circular dependency problem".

The "best", and easiest, solution (which isn't a best practice btw) it to inject the container and then get the templating service.

See: https://github.com/symfony/symfony/issues/2347

But if we don't inject in the constructor but use assigment instead, we don't have the problem.